[api] Use File.join to join root_path with URL rest in API tests.
Using + operator to join root_path and rest of URL starting with '/'
leads to a string that starts with '//' if root_path is '/'.
This is undesirable because part of URL following leading '//'
can be interpreted as server name, not path component.
Using File.join is much safer because it produces correct path
without double slashes.

Signed-off-by: Oleg Girko <ol@infoserver.lv>