[frontend] Drop halt_callback_chains_on_return_false
In Rails 5 before_* / after_* callbacks that return false do not halt
the callback chain anymore. We were keeping this behaviour alive via
Rails's new_framework_defaults initializer.
But this got deprecated in 5.1. So as preparation for 5.2 we had to
verify that we don't need this anymore and finally drop it.
Which is done with this commit;-)

And because of this change we don't have to explicitly return true
in these callbacks anymore.