[frontend] Rails 5.2: Add a prefix to kiwi preference enum
With the recent rails update we started to get an error message:

'You tried to define an enum named "type_image" on the model "Kiwi::Preference",
 but this will generate a class method "split", which is already defined by
 ActiveRecord::Relation. (ArgumentError)
'

The solution to that is to define a prefix for that enumaration. This
prefix will be added to any helper method generated by the enum library,
eg. my_preference.image_type_split? instead of my_preference.split?.