[frontend] Rails 5.2: Update callback in bs request model
We were using a combination of after_update callback and
ActiveModel::Dirty to track changes of bs request states.

After upgrading to Rails 5.2 this stopped to work and we saw
some test failures.

The reason for that was that the ActiveModel::Dirty helper
stopped registering attribute changes after the update was
made (but before the save).
Which seems to make a lot of sense, but it caused this code
to always match:

`
  return if state_was.to_s == state.to_s
`

Using the before_update callback instead solved this.