[frontend] Properly remove annotate gem
This is a follow up on 499905fac8d87e23218dd6a43262859fbfccba94.
We remove annotate until we switch to bundle-gems service or our gem packaging scripts have a way
to deal with conflicting executeables.

As the previous commit did not remove annotate properly we reverted
it and do it in this commit.
We can revert this whole commit when we want to enable annotate again.