[frontend] Require some dependencies explicitly.
Loading some modules started breaking for some reason in Fedora 32
using bundler_ext with "Undefined constant" error for some class names
that are defined in a different module.

This results in immediate error when eager loading is used
(if class name causing error is defined in a module eager loaded later)
or delayed error is eager loading is not used.

The cause of this problem is unclear. Probably, this could be because
Fedora 32 switched from bundler 1 to bundler 2 (2.1.4 to be exact).

Anyway, this change adds explicit requires to work around this problem.

Signed-off-by: Oleg Girko <ol@infoserver.lv>