[api] Pass keyword args properly for Ruby 3.0.
In Ruby 3.0, *args in argument list captutes only positional arguments,
but not keyword ones.
Hence, **kwargs should be added to capture keyword arguments to pass
them correctly to wrapped function.

However, this can't be done unconditionally because it has unintended
consequences in Ruby 2 when passing something convertible to hash
(that has to_hash method) as the last argement to any wrapped function:
it is converted to hash and passed as keyword arguments.

This leads to miserable failure when passing
ActionController::Parameters, for example: its to_hash method raises
ActionController::UnfilteredParameters exception if these parameters
are not marked explicitly as safe.

The only solution I see is to define wrapper different way for
Ruby 3 and Ruby 2: with and without **kwargs argument respectively.

Signed-off-by: Oleg Girko <ol@infoserver.lv>