[dist] Don't require obs-common from obs-worker.
The only file in obs-common needed by obs-worker is
"/etc/sysconfig/obs-server", so it makes more sense to add this file
to obs-worker as well.

There are no other files in obs-common that are needed in obs-worker.

The only reason why it was required before were obsrun user and group,
but now they are provided by sepatare packages.

Signed-off-by: Oleg Girko <ol@infoserver.lv>