[api] Patch strscan gem version in Gemfile.lock file.
Newer versions of Ruby have built-in strscan gem that has version
different than the one bundled in this project.

Insisting on the version of the bundled gem in Gemfile.lock leads to a
version conflict.

As a workaround, patching Gemfile.lock is needed, at least until bettter
gem version conflict resolution lands in Ruby stable version.

See
https://github.com/rubygems/rubygems/pull/5529
https://github.com/rubygems/rubygems/pull/5535
for details.

Signed-off-by: Oleg Girko <ol@infoserver.lv>